Yaml Configuration refactored
